third_party/WebKit/LayoutTests/editing/deleting/display-table-expected.txt

This test verifies that we are able to successfully delete the first and the last characters of a contenteditable div with display: table.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('tableDiv').textContent is "BCD"
PASS document.getElementById('tableDiv').textContent is "ABC"
PASS successfullyParsed is true

TEST COMPLETE
